backend/requirements.txt
# Alpaca Trading API
alpaca-py==0.21.0

# Data Analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# API Framework
flask==3.0.0
flask-cors==4.0.0

# Telegram Bot
python-telegram-bot==20.7

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Date/Time Utilities
python-dateutil==2.8.2

# JSON handling
jsonschema==4.20.0

# Async support
aiohttp==3.9.1
